requirements.txt
numpy==1.14.5
matplotlib==3.0.0
scipy==1.2.0
Pillow==5.3.0
tqdm==4.28.1
opencv-python==3.4.4.19
tensorflow-gpu==1.10